refactor: :recycle: pages to pages folder
