+++ assets/src/pages/home/page.tsx
@@ -1,8 +1,8 @@
-import React, { useEffect } from 'react';
-import { useStore } from '../../main/context-provider';
+import { useEffect } from 'react';
+import { useStore } from '~/main/context-provider';
 import { observer } from 'mobx-react-lite';
-import { SupervisorBlocks } from '../components/supervisor-blocks';
-import { LoadingBlock } from '../components/loading-block';
+import { SupervisorBlocks } from '~/landing/components/supervisor-blocks';
+import { LoadingBlock } from '~/landing/components/loading-block';
 
 export const LandingPage = observer(() => {
   const { landingStore } = useStore();
