+++ assets/src/pages/not-found/index.ts
@@ -1,0 +1,1 @@
+export { NotFoundPage } from './page';
